+++ ann_benchmarks/algorithms/aliyun_elasticsearch.py
@@ -1,7 +1,9 @@
 from __future__ import absolute_import
+import asyncio
 import uuid
-from elasticsearch import Elasticsearch
-from elasticsearch.helpers import bulk
+import time
+from elasticsearch import AsyncElasticsearch
+from elasticsearch.helpers import async_bulk
 from ann_benchmarks.algorithms.base import BaseANN
 
 
@@ -14,14 +16,21 @@ class AliESHNSW(BaseANN):
         self._ef = None
         self._field = "vec"
         # self._es = Elasticsearch([ip], port=port)
-        self._es = Elasticsearch("http://es-cn-oew1t7131000y8424.public.elasticsearch.aliyuncs.com:9200",
-                                 http_auth=("******", "********"), max_retries=5, retry_on_timeou=True, timeout=30)
+        self._es = AsyncElasticsearch("http://es-cn-oew1t7131000y8424.elasticsearch.aliyuncs.com:9200",
+                                      http_auth=("******", "********"), max_retries=5, retry_on_timeout=True, timeout=30)
+        self._loop = asyncio.get_event_loop()
 
     def get_memory_usage(self):
-        if not self._es.indices.exists(index=self._index_name):
+        # a, b = loop.run_until_complete(asyncio.gather(*tasks))
+        # exists = self._loop.run_until_complete(asyncio.gather(self._es.indices.exists(index=self._index_name)))
+        exists = self._loop.run_until_complete(self._es.indices.exists(index=self._index_name))
+        if not exists:
             return 0
+        # if not self._es.indices.exists(index=self._index_name):
+        #     return 0
 
-        stats = self._es.indices.stats(index=self._index_name)
+        stats = self._loop.run_until_complete(self._es.indices.stats(index=self._index_name))
+        # stats = self._es.indices.stats(index=self._index_name)
         return int(stats["_all"]["total"]["store"]["size_in_bytes"]) / 1024
 
     def fit(self, X):
@@ -48,23 +57,29 @@ class AliESHNSW(BaseANN):
             }
         }
 
-        if self._es.indices.exists(index=self._index_name):
-            self._es.indices.delete(index=self._index_name)
+        exist = self._loop.run_until_complete(self._es.indices.exists(index=self._index_name))
+        if exist:
+            self._loop.run_until_complete(self._es.indices.delete(index=self._index_name))
         # create index mappings
-        self._es.indices.create(index=self._index_name, body=_index_body)
-        bulk_records = []
-        for i, vec in enumerate(X.tolist()):
-            # print("i: ", i, ", vec: ", vec)
-            doc_template = {
-                "_index": self._index_name,
-                "_id": i,
-                "_source": {
-                    self._field: vec,
+        self._loop.run_until_complete(self._es.indices.create(index=self._index_name, body=_index_body))
+
+        # bulk_records = []
+        async def gen_action():
+            for i, vec in enumerate(X.tolist()):
+                # print("i: ", i, ", vec: ", vec)
+                yield {
+                    "_index": self._index_name,
+                    "_id": i,
+                    "_source": {
+                        self._field: vec,
+                    }
                 }
-            }
-            bulk_records.append(doc_template)
-        success, _ = bulk(self._es, bulk_records, index=self._index_name, raise_on_error=True)
-        self._es.indices.refresh(index=self._index_name)  # refresh to update index
+                # bulk_records.append(doc_template)
+
+        success, failed = self._loop.run_until_complete(
+            async_bulk(self._es, gen_action(), index=self._index_name, raise_on_error=True))
+        # success, _ = bulk(self._es, bulk_records, index=self._index_name, raise_on_error=True)
+        self._loop.run_until_complete(self._es.indices.refresh(index=self._index_name))  # refresh to update index
 
     def set_query_arguments(self, ef):
         self._ef = ef
@@ -83,15 +98,31 @@ class AliESHNSW(BaseANN):
             "_source": False
         }
 
-        top_n_results = self._es.search(index=self._index_name, body=query_params)
+        return self._es.search(index=self._index_name, body=query_params)
         # print("top_n_results: ", top_n_results)
-        top_n_results = [int(doc['_id']) for doc in top_n_results['hits']['hits']]
-        return top_n_results
+        # top_n_results = [int(doc['_id']) for doc in top_n_results['hits']['hits']]
+        # return top_n_results
+
+    def handle_query_list_result(self, query_list):
+        t0 = time.time()
+        ql = (future for _, _,  future in query_list)
+        results = self._loop.run_until_complete(asyncio.gather(*ql))
+        handled_result = []
+
+        for q, result in zip(query_list, results):
+            total, v, _ = q
+            result_ids = [int(doc['_id']) for doc in result['hits']['hits']]
+            handled_result.append([total, v, result_ids])
+
+        return time.time() - t0, handled_result
 
     def done(self):
-        if self._es.indices.exists(index=self._index_name):
+        exists = self._loop.run_until_complete(self._es.indices.exists(index=self._index_name))
+        if exists:
             print("delete index...")
-            self._es.indices.delete(index=self._index_name)
+            self._loop.run_until_complete(self._es.indices.delete(index=self._index_name))
+
+        self._loop.run_until_complete(self._es.close())
 
     def __str__(self):
-        return ("Elasticsearch(%s, %s)" % ("euclidean", self._index_name))
+        return "Elasticsearch({}, param: {}, search param: {})".format("euclidean", self._method_param, self._ef)
